requirements.txt
# Core dependencies
PyMuPDF>=1.23.0
pillow>=10.0.0
python-dotenv==1.0.1

# API clients
google-generativeai==0.8.2
openai==1.51.2
google-cloud-texttospeech==2.17.2

# Video processing
moviepy==1.0.3

# Additional dependencies that may be required
imageio==2.35.1
imageio-ffmpeg==0.5.1 